refactored the application code
