+++ src/main/java/com/gideon/contact_manager/application/dto/UserResponse.java
@@ -1,13 +1,15 @@
 package com.gideon.contact_manager.application.dto;
 
 import com.gideon.contact_manager.domain.model.ContactGroup;
+import lombok.Data;
 
+@Data
 public class UserResponse {
     private String firstName;
     private String lastName;
     private String email;
-    private String phoneNumber;
-    private String imageUrl;
-    private String address;
-    private ContactGroup group;
+//    private String phoneNumber;
+//    private String imageUrl;
+//    private String address;
+//    private ContactGroup group;
 }
