+++ src/main/java/com/gideon/contact_manager/application/features/user/commands/create/CreateUserCommand.java
@@ -1,0 +1,28 @@
+package com.gideon.contact_manager.application.features.user.commands.create;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Temporal;
+import jakarta.persistence.TemporalType;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@Builder
+public class CreateUserCommand {
+    private String firstName;
+    private String lastName;
+    private String email;
+    private Date createdOn;
+    private String createdBy;
+    private Date modifiedOn;
+    private String modifiedBy;
+}
