+++ src/main/java/com/gideon/contact_manager/application/mapper/UserMapper.java
@@ -1,16 +1,17 @@
 package com.gideon.contact_manager.application.mapper;
 
-import com.gideon.contact_manager.application.dto.CreateUserRequest;
 import com.gideon.contact_manager.application.dto.UserResponse;
+import com.gideon.contact_manager.application.features.user.commands.create.CreateUserCommand;
 import com.gideon.contact_manager.domain.model.User;
 import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
 import org.springframework.stereotype.Component;
 
-@Mapper(componentModel = "spring")
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface UserMapper {
 //    UserMapper INSTANCE = Mappers.getMapper(UserMapper.class);
 
     UserResponse toDTO(User user);
-    User toEntity(CreateUserRequest createUserRequest);
+    User toEntity(CreateUserCommand createUserCommand);
 }
