+++ src/main/java/com/gideon/contact_manager/application/usecase/user/UserServiceImpl.java
@@ -1,15 +1,19 @@
 package com.gideon.contact_manager.application.usecase.user;
 
-import com.gideon.contact_manager.application.dto.CreateUserRequest;
+import com.gideon.contact_manager.application.features.user.commands.create.CreateUserCommand;
+import com.gideon.contact_manager.presentation.apimodels.CreateUserRequest;
 import com.gideon.contact_manager.application.dto.UserResponse;
 import com.gideon.contact_manager.application.mapper.UserMapper;
 import com.gideon.contact_manager.application.service.user.UserService;
+import com.gideon.contact_manager.domain.model.User;
 import com.gideon.contact_manager.infrastructure.persistence.JpaUserRepository;
 import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,9 +32,17 @@ public class UserServiceImpl implements UserService {
 //    }
     @Override
     public UserResponse createUser(CreateUserRequest request) {
-        var newUser = userMapper.toEntity(request);
+        CreateUserCommand cmd = CreateUserCommand
+                .builder()
+                .firstName(request.getFirstName())
+                .lastName(request.getLastName())
+                .email(request.getEmail())
+                .createdBy(request.getFirstName())
+                .createdOn(Date.from(Instant.now()))
+                .build();
+
+        var newUser = userMapper.toEntity(cmd);
         var createdUser = JpaUserRepository.save(newUser);
-//        return UserMapper.INSTANCE.toDTO(createdUser);
         return userMapper.toDTO(createdUser);
     }
